backend/requirements.txt
# Python dependencies for Store343 Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pillow==10.2.0
google-cloud-vision==3.5.0
google-auth==2.27.0
requests==2.31.0
